+++ rando/trekking/tests/__init__.py
@@ -5,14 +5,15 @@ import re
 
 from mock import patch
 from easydict import EasyDict as edict
+from casper.tests import CasperTestCase
 
 from django.test import SimpleTestCase
 from django.conf import settings
 
 
-from .templatetags.trekking_tags import overridable
-from .management.commands.sync_trekking import mkdir_p
-from .models import Trek
+from rando.trekking.templatetags.trekking_tags import overridable
+from rando.trekking.management.commands.sync_trekking import mkdir_p
+from rando.trekking.models import Trek
 
 
 class MkdirTest(SimpleTestCase):
@@ -99,3 +100,25 @@ class TrekFulltextTest(SimpleTestCase):
     def test_duplicate_words_are_removed(self):
         count = len(list(re.finditer('like', self.fulltext)))
         self.assertEqual(count, 1)
+
+
+class NavigationTest(CasperTestCase):
+
+    def setUp(self):
+        basepath = os.path.join(settings.FLATPAGES_ROOT, 'fr')
+        try:
+            os.makedirs(basepath)
+        except OSError:
+            pass
+        self.flatpage = os.path.join(basepath, 'flatpage.html')
+        open(self.flatpage, 'w').write('<h1>test</h1>')
+
+    def tearDown(self):
+        os.remove(self.flatpage)
+
+    def _get_tests_file(self, name):
+        return os.path.join(settings.PROJECT_PATH, 'trekking', 'tests', name)
+
+    def test_popup(self):
+        with self.settings(POPUP_HOME_ENABLED=True):
+            self.assertTrue(self.casper(self._get_tests_file('test_popup.js')))