+++ CHANGES.rst
@@ -19,6 +19,9 @@ CHANGELOG
 
 * All national park branding has been replaced by generic Geotrek material.
   See *Customization* paragraph about images.
+* Search popup is not shown outside homepage anymore. A click on main page title
+  brings it, and the home button in navigation does not.
+* No longer compatible with Django 1.4.
 
 **New features**